app/requirements.txt
redis
streamlit==1.18.0
altair
python-dotenv
langchain
sentence-transformers
tenacity
pandas
numpy
openai
tiktoken
